Modified Driver to take motor co-ordinates in radians, and to have max and min angles
There were a few design decisions that had to be made here. Because the mcbmini classes don't seem to have any functionality to access max_ticks, min_ticks, and ticks_per_unit, I had to create a Map for each board which maps those keys to values. I named this ArrayList of Maps extraParams. It would have been more elegant to have a way to access those params from the board objects.

Another decision was how to parse the config file for those params. parseConfigFile was already doing one pass through the config file. However, because I don't have access to the .java files for XMLUtils or XMLResults, it's not clear what the interface is to access the required fields. So, I had to reparse the config file with XML tools from the JAVA library.

Radians were converted to ticks by multiplying by the TICKS_PER_UNIT parameter in the config file for the given motor. This value should perhaps be offset by a constant.

Lower and upper limits for the motor ticks were added.